# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "一款面向小型团队的任务进度共享软件",
        "title": "Team",
        "termsOfService": "http://swagger.io/terrms",
        "contact": {},
        "version": "1.0.0"
    },
    "host": "localhost:9918",
    "paths": {
...
        "/team/project/:project_id": {
            "get": {
                "description": "\"单击编辑查看项目信息\"",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "project"
                ],
                "summary": "\"查看项目\"",
                "parameters": [
                    {
                        "type": "string",
                        "description": "token",
                        "name": "token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "project_id",
                        "name": "project_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "获取成功"
                    },
                    "401": {
                        "description": "身份验证失败"
                    },
                    "404": {
                        "description": "获取失败"
                    }
                }
            },
            "put": {
                "description": "\"创建人修改一个项目\"",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "project"
                ],
                "summary": "\"修改项目\"",
                "parameters": [
                    {
                        "type": "string",
                        "description": "token",
                        "name": "token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "project_id",
                        "name": "project_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "修改成功"
                    },
                    "400": {
                        "description": "删除失败"
                    },
                    "401": {
                        "description": "身份验证失败"
                    }
                }
            },
            "delete": {
                "description": "\"删除一个项目\"",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "project"
                ],
                "summary": "\"删除项目\"",
                "parameters": [
                    {
                        "type": "string",
                        "description": "token",
                        "name": "token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "project_id",
                        "name": "project_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    },
                    "400": {
                        "description": "删除失败"
                    },
                    "401": {
                        "description": "身份验证失败"
                    }
                }
            }
        },
...
    },
...
}